renamed previous burndown chart to product burndown, created a new burndown chart for each sprint which is included in the sprint history template, sprint history of tasks done is also included in that page, added developers working for that project to help with ideal remaining during sprint burndown chart calculation, estimated sprints required remained for product burndown chart

+++ src/main/java/gr/university/thesis/dto/BurnDownChartData.java
@@ -23,8 +23,10 @@ public class BurnDownChartData {
     /**
      * ideal_burn is the array that contains the ideal burn of effort based on the project total effort divided by
      * the estimated sprints needed for that project to finish, for example [100(total project effort), 86, ...]
+     * it is represented as a double, as the division can result in a double, and if it not properly handled, it can
+     * be inaccurate
      */
-    int[] ideal_burn;
+    double[] ideal_burn;
 
     /**
      * actual_burn is the array that contains the actual burn of effort (velocity) that was done for each finished
